requirements/dev.txt
-r common.txt
django-debug-toolbar==1.5

